backend/pyproject.toml
[tool.poetry]
name = "supplifit"
version = "0.1.0"
description = "SuppliFit - Um Gympass dos suplementos"
authors = ["sumioshi <sumioshi@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.7"
djangorestframework = "3.15.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
django-filter = "^24.1"
scikit-learn = "^1.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 